requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy
annotated-types     
anyio               
asgiref             
click               
colorama            
customtkinter      
darkdetect         
Django             
djangorestframework 
environs           
fastapi            
h11                 
idna               
joblib             
marshmallow       
numpy              
packaging        
pillow             
pip                 
pydantic            
pydantic_core       
python-dotenv     
python-multipart    
scikit-learn        
scipy              
setuptools         
sniffio           
sqlparse           
starlette           
threadpoolctl       
typing_extensions  
tzdata             
uvicorn            
wheel               